requirements.txt
Flask==2.0.3
Flask_Login==0.5.0
Flask_SQLAlchemy==2.5.1
h5py==3.6.0
keras==2.8.0
numpy==1.21.4
Pillow==9.0.1
SQLAlchemy==1.3.11
tensorflow-cpu==2.8.0
Werkzeug==2.0.3
gunicorn==19.0.0